+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+#### Breaking changes
+
+- **Breaking:** The `TransientSource` is now an opaque type. It provides API methods for removing or
+  replacing the wrapped source. This mitigates a potential leak of registration data if the
+  TransientSource is replaced by direct assignment in a parent source.
+
 ## 0.10.1 -- 2022-06-20
 
 #### Additions